Create a base code with customized functions to validate all fields.
Create an array with constants for every field. Create a CheckValidity function to check the validity of each field according to the HTML validation. Create a ShowError function that creates an Error message constant for each field and then checks if each field complies with all the requirements in HTML.